+++ src/views/ExampleView.vue
@@ -1,37 +1,19 @@
+<template>
+  <main class="container">
+    <h1>
+      <div class="align-middle">
+        <ScannerSvg class="icon" />
+        <span>Look Scanned</span>
+      </div>
+    </h1>
+    <MarkdownContent />
+  </main>
+</template>
+
 <script setup lang="ts">
-import zhCNMarkdown from '../locales/examples/markdowns/zh-CN.md?raw'
-import { useRoute } from 'vue-router'
-import { parse } from 'marked'
-import DOMPurify from 'dompurify'
-import { computed } from 'vue'
 import ScannerSvg from '../components/ScannerSvg.vue'
-import { computedAsync } from '@vueuse/core'
-
-const route = useRoute()
-const lang = route.params.lang
-
-const fonts = {
-  'zh-CN': {
-    'font-family': 'Noto Serif SC',
-    'font-style': 'normal',
-  },
-}
-
-const markdownHTML = computedAsync(async () => {
-  console.log(lang)
-  // get markdown
-  const markdown = zhCNMarkdown
-  return DOMPurify.sanitize(await parse(markdown))
-})
-
-const style = computed(() => {
-  return fonts?.[lang as keyof typeof fonts] ?? {}
-})
+import MarkdownContent from '@/components/MarkdownContent.vue'
 </script>
-
-<style>
-@import url('https://fonts.googleapis.com/css2?family=Noto+Serif+SC:wght@200..900&display=swap');
-</style>
 
 <style scoped>
 .icon {
@@ -58,10 +40,6 @@ h1 {
   margin: 0 auto;
   padding: 1in;
 }
-
-.markdown {
-  font-size: 1.5rem;
-}
 </style>
 
 <style>
@@ -70,15 +48,3 @@ h1 {
   margin: 0cm; /* Default print margin */
 }
 </style>
-
-<template>
-  <main class="container">
-    <h1>
-      <div class="align-middle">
-        <ScannerSvg class="icon" />
-        <span>Look Scanned</span>
-      </div>
-    </h1>
-    <div v-html="markdownHTML" :style="style" class="markdown"></div>
-  </main>
-</template>
